DjangoRestApi/DjangoRestFramework/src/ManagementProject/requirement.txt
chardet==2.3.0
cryptography==1.7.1
cupshelpers==1.0
Django==2.2.7
django-admin-sortable==2.0.10
djangorestframework==3.10.3
idna==2.2
keyring==10.1
keyrings.alt==1.3
onboard==1.4.1
pexpect==4.2.1
Pillow==4.0.0
ptyprocess==0.5.1
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pydevd-pycharm==193.5233.15
pygobject==3.22.0
pysmbc==1.0.15.6
pytz==2019.3
pyxdg==0.25
reportlab==3.3.0
requests==2.12.4
SecretStorage==2.3.1
six==1.10.0
sqlparse==0.3.0
urllib3==1.19.1
virtualenv==15.1.0
yolk3k==0.9
